move nextauth to app folder
