+++ apps/web/app/onboarding/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth";
 import Onboarding from "./Onboarding";
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 
 export default async function OnboardingPage() {
   const session = await getServerSession(authOptions);
