+++ apps/web/lib/api/apiHelper.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { authOptions } from "@/app/api/auth/[...nextauth]/authOptions";
 import { prisma } from "@formbricks/database";
 import { createHash } from "crypto";
 import { NextApiRequest, NextApiResponse } from "next";
